+++ benchmarks/packages/Sudoku.jl
@@ -1,0 +1,10 @@
+
+module SudokuBenchmarks
+
+using Sudoku
+
+sudoku_constraints() = sudoku(zeros(Int, 9, 9))
+
+const functions = [sudoku_constraints]
+
+end # SudokuBenchmarks